requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
pydantic==2.10.4
cryptography==44.0.0
apscheduler==3.10.4
playwright>=1.41.0
playwright-stealth>=1.0.6
setuptools>=69.0
python-dotenv==1.0.1
anthropic>=0.40.0
loguru>=0.7.0
httpx>=0.27.0
